--- mcp_tools/cloud_container/cloud_and_container_security.py
@@ -1,105 +1,0 @@
-# mcp_tools/cloud_and_container_security.py
-
-from typing import Dict, Any
-
-def register_cloud_and_container_security_tools(mcp, hexstrike_client, logger):
-
-    @mcp.tool()
-    def falco_runtime_monitoring(config_file: str = "/etc/falco/falco.yaml",
-                                rules_file: str = "", output_format: str = "json",
-                                duration: int = 60, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Falco for runtime security monitoring.
-
-        Args:
-            config_file: Falco configuration file
-            rules_file: Custom rules file
-            output_format: Output format (json, text)
-            duration: Monitoring duration in seconds
-            additional_args: Additional Falco arguments
-
-        Returns:
-            Runtime security monitoring results
-        """
-        data = {
-            "config_file": config_file,
-            "rules_file": rules_file,
-            "output_format": output_format,
-            "duration": duration,
-            "additional_args": additional_args
-        }
-        logger.info(f"🛡️  Starting Falco runtime monitoring for {duration}s")
-        result = hexstrike_client.safe_post("api/tools/falco", data)
-        if result.get("success"):
-            logger.info(f"✅ Falco monitoring completed")
-        else:
-            logger.error(f"❌ Falco monitoring failed")
-        return result
-
-    @mcp.tool()
-    def checkov_iac_scan(directory: str = ".", framework: str = "", check: str = "",
-                        skip_check: str = "", output_format: str = "json",
-                        additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Checkov for infrastructure as code security scanning.
-
-        Args:
-            directory: Directory to scan
-            framework: Framework to scan (terraform, cloudformation, kubernetes, etc.)
-            check: Specific check to run
-            skip_check: Check to skip
-            output_format: Output format (json, yaml, cli)
-            additional_args: Additional Checkov arguments
-
-        Returns:
-            Infrastructure as code security scanning results
-        """
-        data = {
-            "directory": directory,
-            "framework": framework,
-            "check": check,
-            "skip_check": skip_check,
-            "output_format": output_format,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Checkov IaC scan: {directory}")
-        result = hexstrike_client.safe_post("api/tools/checkov", data)
-        if result.get("success"):
-            logger.info(f"✅ Checkov scan completed")
-        else:
-            logger.error(f"❌ Checkov scan failed")
-        return result
-
-    @mcp.tool()
-    def terrascan_iac_scan(scan_type: str = "all", iac_dir: str = ".",
-                          policy_type: str = "", output_format: str = "json",
-                          severity: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Terrascan for infrastructure as code security scanning.
-
-        Args:
-            scan_type: Type of scan (all, terraform, k8s, etc.)
-            iac_dir: Infrastructure as code directory
-            policy_type: Policy type to use
-            output_format: Output format (json, yaml, xml)
-            severity: Severity filter (high, medium, low)
-            additional_args: Additional Terrascan arguments
-
-        Returns:
-            Infrastructure as code security scanning results
-        """
-        data = {
-            "scan_type": scan_type,
-            "iac_dir": iac_dir,
-            "policy_type": policy_type,
-            "output_format": output_format,
-            "severity": severity,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Terrascan IaC scan: {iac_dir}")
-        result = hexstrike_client.safe_post("api/tools/terrascan", data)
-        if result.get("success"):
-            logger.info(f"✅ Terrascan scan completed")
-        else:
-            logger.error(f"❌ Terrascan scan failed")
-        return result
