+++ mcp_tools/iac_scan/terrascan.py
@@ -1,0 +1,38 @@
+# mcp_tools/iac_scan/terrascan.py
+
+from typing import Dict, Any
+
+def register_terrascan_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def terrascan_iac_scan(scan_type: str = "all", iac_dir: str = ".",
+                          policy_type: str = "", output_format: str = "json",
+                          severity: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Terrascan for infrastructure as code security scanning.
+
+        Args:
+            scan_type: Type of scan (all, terraform, k8s, etc.)
+            iac_dir: Infrastructure as code directory
+            policy_type: Policy type to use
+            output_format: Output format (json, yaml, xml)
+            severity: Severity filter (high, medium, low)
+            additional_args: Additional Terrascan arguments
+
+        Returns:
+            Infrastructure as code security scanning results
+        """
+        data = {
+            "scan_type": scan_type,
+            "iac_dir": iac_dir,
+            "policy_type": policy_type,
+            "output_format": output_format,
+            "severity": severity,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Terrascan IaC scan: {iac_dir}")
+        result = hexstrike_client.safe_post("api/tools/terrascan", data)
+        if result.get("success"):
+            logger.info(f"✅ Terrascan scan completed")
+        else:
+            logger.error(f"❌ Terrascan scan failed")
+        return result
