+++ mcp_tools/runtime_monitor/falco.py
@@ -1,0 +1,37 @@
+# mcp_tools/runtime_monitor/falco.py
+
+from typing import Dict, Any
+
+def register_falco_runtime_monitoring_tool(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def falco_runtime_monitoring(config_file: str = "/etc/falco/falco.yaml",
+                                rules_file: str = "", output_format: str = "json",
+                                duration: int = 60, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Falco for runtime security monitoring.
+
+        Args:
+            config_file: Falco configuration file
+            rules_file: Custom rules file
+            output_format: Output format (json, text)
+            duration: Monitoring duration in seconds
+            additional_args: Additional Falco arguments
+
+        Returns:
+            Runtime security monitoring results
+        """
+        data = {
+            "config_file": config_file,
+            "rules_file": rules_file,
+            "output_format": output_format,
+            "duration": duration,
+            "additional_args": additional_args
+        }
+        logger.info(f"🛡️  Starting Falco runtime monitoring for {duration}s")
+        result = hexstrike_client.safe_post("api/tools/falco", data)
+        if result.get("success"):
+            logger.info(f"✅ Falco monitoring completed")
+        else:
+            logger.error(f"❌ Falco monitoring failed")
+        return result
